fix: Update package references and handle exceptions more broadly
