+++ src/KSail.Generator/KSail.Generator.csproj
@@ -12,7 +12,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Devantler.KubernetesGenerator.Native" Version="1.2.7" />
+    <PackageReference Include="Devantler.KubernetesGenerator.Native" Version="1.5.0" />
   </ItemGroup>
 
   <ItemGroup>
