+++ src/KSail.Models/KSail.Models.csproj
@@ -13,7 +13,7 @@
 
   <ItemGroup>
     <PackageReference Include="YamlDotNet" Version="16.3.0" />
-    <PackageReference Include="Devantler.K9sCLI" Version="1.1.4" />
+    <PackageReference Include="Devantler.K9sCLI" Version="1.2.1" />
   </ItemGroup>
 
 </Project>
