+++ src/KSail/Commands/Gen/Commands/CertManager/KSailGenCertManagerCertificateCommand.cs
@@ -23,7 +23,7 @@ class KSailGenCertManagerCertificateCommand : Command
           Console.WriteLine($"✚ generating {outputFile}");
           context.ExitCode = await handler.HandleAsync(outputFile, context.GetCancellationToken()).ConfigureAwait(false);
         }
-        catch (OperationCanceledException ex)
+        catch (Exception ex)
         {
           _ = _exceptionHandler.HandleException(ex);
           context.ExitCode = 1;
