+++ src/KSail/Commands/Gen/Commands/CertManager/KSailGenCertManagerClusterIssuerCommand.cs
@@ -24,7 +24,7 @@ class KSailGenCertManagerClusterIssuerCommand : Command
           Console.WriteLine($"✚ generating {outputFile}");
           context.ExitCode = await handler.HandleAsync(outputFile, context.GetCancellationToken()).ConfigureAwait(false);
         }
-        catch (OperationCanceledException ex)
+        catch (Exception ex)
         {
           _ = _exceptionHandler.HandleException(ex);
           context.ExitCode = 1;
