+++ src/KSail/Commands/Gen/Commands/Config/KSailGenConfigKSailCommand.cs
@@ -22,7 +22,7 @@ class KSailGenConfigKSailCommand : Command
           Console.WriteLine($"✚ generating {outputFile}");
           context.ExitCode = await _handler.HandleAsync(outputFile, context.GetCancellationToken()).ConfigureAwait(false);
         }
-        catch (OperationCanceledException ex)
+        catch (Exception ex)
         {
           _ = _exceptionHandler.HandleException(ex);
           context.ExitCode = 1;
