+++ src/KSail/Commands/Gen/Commands/Flux/KSailGenFluxKustomizationCommand.cs
@@ -23,7 +23,7 @@ sealed class KSailGenFluxKustomizationCommand : Command
           Console.WriteLine($"✚ generating {outputFile}");
           context.ExitCode = await _handler.HandleAsync(outputFile, context.GetCancellationToken()).ConfigureAwait(false);
         }
-        catch (OperationCanceledException ex)
+        catch (Exception ex)
         {
           _ = _exceptionHandler.HandleException(ex);
           context.ExitCode = 1;
