+++ src/KSail/Commands/Gen/Commands/Native/KSailGenNativePodDisruptionBudgetCommand.cs
@@ -23,7 +23,7 @@ class KSailGenNativePodDisruptionBudgetCommand : Command
           Console.WriteLine($"✚ generating {outputFile}");
           context.ExitCode = await _handler.HandleAsync(outputFile, context.GetCancellationToken()).ConfigureAwait(false);
         }
-        catch (OperationCanceledException ex)
+        catch (Exception ex)
         {
           _ = _exceptionHandler.HandleException(ex);
           context.ExitCode = 1;
