+++ src/KSail/Commands/Gen/Commands/Native/KSailGenNativePriorityClassCommand.cs
@@ -23,7 +23,7 @@ class KSailGenNativePriorityClassCommand : Command
           Console.WriteLine($"✚ generating {outputFile}");
           context.ExitCode = await _handler.HandleAsync(outputFile, context.GetCancellationToken()).ConfigureAwait(false);
         }
-        catch (OperationCanceledException ex)
+        catch (Exception ex)
         {
           _ = _exceptionHandler.HandleException(ex);
           context.ExitCode = 1;
