+++ src/KSail/Commands/Secrets/Commands/KSailSecretsDecryptCommand.cs
@@ -45,7 +45,7 @@ sealed class KSailSecretsDecryptCommand : Command
         context.ExitCode = await handler.HandleAsync(context.GetCancellationToken()).ConfigureAwait(false);
         Console.WriteLine();
       }
-      catch (OperationCanceledException ex)
+      catch (Exception ex)
       {
         _ = _exceptionHandler.HandleException(ex);
         context.ExitCode = 1;
