+++ src/KSail/Commands/Secrets/Commands/KSailSecretsEditCommand.cs
@@ -1,27 +1,31 @@
 using System.CommandLine;
 using Devantler.SecretManager.SOPS.LocalAge;
+using KSail.Commands.Secrets.Arguments;
 using KSail.Commands.Secrets.Handlers;
 using KSail.Options;
 using KSail.Utils;
 
 namespace KSail.Commands.Secrets.Commands;
 
-sealed class KSailSecretsGenerateCommand : Command
+sealed class KSailSecretsEditCommand : Command
 {
   readonly ExceptionHandler _exceptionHandler = new();
+  readonly PathArgument _pathArgument = new("The path to the file to edit.") { Arity = ArgumentArity.ExactlyOne };
   readonly ProjectSecretManagerOption _projectSecretManagerOption = new() { Arity = ArgumentArity.ZeroOrOne };
 
-  internal KSailSecretsGenerateCommand() : base("gen", "Generate a new encryption key")
+  internal KSailSecretsEditCommand() : base("edit", "Edit an encrypted file")
   {
-    AddOption(_projectSecretManagerOption);
+    AddArgument(_pathArgument);
+    AddOptions();
     this.SetHandler(async (context) =>
     {
       try
       {
         var config = await KSailClusterConfigLoader.LoadAsync().ConfigureAwait(false);
         config.UpdateConfig("Spec.Project.SecretManager", context.ParseResult.GetValueForOption(_projectSecretManagerOption));
+        string path = context.ParseResult.GetValueForArgument(_pathArgument);
         var cancellationToken = context.GetCancellationToken();
-        KSailSecretsGenerateCommandHandler handler;
+        KSailSecretsEditCommandHandler handler;
         switch (config.Spec.Project.SecretManager)
         {
           default:
@@ -30,11 +34,9 @@ sealed class KSailSecretsGenerateCommand : Command
             context.ExitCode = 1;
             return;
           case Models.Project.KSailSecretManager.SOPS:
-            handler = new KSailSecretsGenerateCommandHandler(new SOPSLocalAgeSecretManager());
+            handler = new KSailSecretsEditCommandHandler(config, path, new SOPSLocalAgeSecretManager());
             break;
         }
-
-        Console.WriteLine($"🔑 Generating a new encryption key with '{config.Spec.Project.SecretManager}'");
         context.ExitCode = await handler.HandleAsync(context.GetCancellationToken()).ConfigureAwait(false);
         Console.WriteLine();
       }
@@ -45,6 +47,6 @@ sealed class KSailSecretsGenerateCommand : Command
       }
     });
   }
+
+  void AddOptions() => AddOption(_projectSecretManagerOption);
 }
-
-
