+++ src/KSail/Commands/Secrets/Commands/KSailSecretsEncryptCommand.cs
@@ -47,7 +47,7 @@ sealed class KSailSecretsEncryptCommand : Command
         context.ExitCode = await handler.HandleAsync(context.GetCancellationToken()).ConfigureAwait(false);
         Console.WriteLine();
       }
-      catch (OperationCanceledException ex)
+      catch (Exception ex)
       {
         _ = _exceptionHandler.HandleException(ex);
         context.ExitCode = 1;
