+++ src/KSail/Commands/Secrets/Commands/KSailSecretsExportCommand.cs
@@ -56,7 +56,7 @@ sealed class KSailSecretsExportCommand : Command
         context.ExitCode = await handler.HandleAsync(context.GetCancellationToken()).ConfigureAwait(false);
         Console.WriteLine();
       }
-      catch (OperationCanceledException ex)
+      catch (Exception ex)
       {
         _ = _exceptionHandler.HandleException(ex);
         context.ExitCode = 1;
