+++ src/KSail/Commands/Secrets/Commands/KSailSecretsListCommand.cs
@@ -43,7 +43,7 @@ sealed class KSailSecretsListCommand : Command
         context.ExitCode = await handler.HandleAsync(context.GetCancellationToken()).ConfigureAwait(false) ? 0 : 1;
         Console.WriteLine();
       }
-      catch (OperationCanceledException ex)
+      catch (Exception ex)
       {
         _ = _exceptionHandler.HandleException(ex);
         context.ExitCode = 1;
