+++ src/KSail/Commands/Secrets/KSailSecretsCommand.cs
@@ -19,6 +19,8 @@ sealed class KSailSecretsCommand : Command
   {
     AddCommand(new KSailSecretsEncryptCommand());
     AddCommand(new KSailSecretsDecryptCommand());
+    // TODO: Include `ksail secrets edit` command when pseudo-terminal support is added to CLIWrap. See https://github.com/Tyrrrz/CliWrap/issues/225.
+    //AddCommand(new KSailSecretsEditCommand());
     AddCommand(new KSailSecretsGenerateCommand());
     AddCommand(new KSailSecretsDeleteCommand());
     AddCommand(new KSailSecretsListCommand());
