+++ src/KSail/Commands/Start/KSailStartCommand.cs
@@ -2,7 +2,6 @@ using System.CommandLine;
 using KSail.Commands.Start.Handlers;
 using KSail.Options;
 using KSail.Utils;
-using YamlDotNet.Core;
 
 namespace KSail.Commands.Start;
 
@@ -38,17 +37,7 @@ sealed class KSailStartCommand : Command
           throw new KSailException("Cluster could not be started");
         }
       }
-      catch (YamlException ex)
-      {
-        _ = _exceptionHandler.HandleException(ex);
-        context.ExitCode = 1;
-      }
-      catch (KSailException ex)
-      {
-        _ = _exceptionHandler.HandleException(ex);
-        context.ExitCode = 1;
-      }
-      catch (OperationCanceledException ex)
+      catch (Exception ex)
       {
         _ = _exceptionHandler.HandleException(ex);
         context.ExitCode = 1;
