+++ src/KSail/Commands/Stop/KSailStopCommand.cs
@@ -2,7 +2,6 @@ using System.CommandLine;
 using KSail.Commands.Stop.Handlers;
 using KSail.Options;
 using KSail.Utils;
-using YamlDotNet.Core;
 
 namespace KSail.Commands.Stop;
 
@@ -38,18 +37,9 @@ sealed class KSailStopCommand : Command
           throw new KSailException("Cluster could not be stopped");
         }
       }
-      catch (YamlException ex)
+      catch (Exception ex)
       {
         _ = _exceptionHandler.HandleException(ex);
-        context.ExitCode = 1;
-      }
-      catch (KSailException ex)
-      {
-        _ = _exceptionHandler.HandleException(ex);
-        context.ExitCode = 1;
-      }
-      catch (OperationCanceledException)
-      {
         context.ExitCode = 1;
       }
     });
