+++ src/KSail/KSail.csproj
@@ -21,17 +21,17 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Devantler.ContainerEngineProvisioner.Docker" Version="1.1.1" />
-    <PackageReference Include="Devantler.KubernetesGenerator.CertManager" Version="1.4.1" />
-    <PackageReference Include="Devantler.KubernetesGenerator.Flux" Version="1.4.1" />
-    <PackageReference Include="Devantler.KubernetesGenerator.K3d" Version="1.4.1" />
-    <PackageReference Include="Devantler.KubernetesGenerator.Kind" Version="1.4.1" />
+    <PackageReference Include="Devantler.ContainerEngineProvisioner.Docker" Version="1.1.3" />
+    <PackageReference Include="Devantler.KubernetesGenerator.CertManager" Version="1.5.0" />
+    <PackageReference Include="Devantler.KubernetesGenerator.Flux" Version="1.5.0" />
+    <PackageReference Include="Devantler.KubernetesGenerator.K3d" Version="1.5.0" />
+    <PackageReference Include="Devantler.KubernetesGenerator.Kind" Version="1.5.0" />
     <PackageReference Include="Devantler.KubernetesProvisioner.Cluster.K3d" Version="1.2.12" />
     <PackageReference Include="Devantler.KubernetesProvisioner.Cluster.Kind" Version="1.2.12" />
     <PackageReference Include="Devantler.KubernetesProvisioner.GitOps.Flux" Version="1.2.12" />
     <PackageReference Include="Devantler.KubernetesValidator.ClientSide.Schemas" Version="1.0.24" />
     <PackageReference Include="Devantler.KubernetesValidator.ClientSide.YamlSyntax" Version="1.0.24" />
-    <PackageReference Include="Devantler.SecretManager.SOPS.LocalAge" Version="1.2.4" />
+    <PackageReference Include="Devantler.SecretManager.SOPS.LocalAge" Version="1.3.0" />
     <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
   </ItemGroup>
